feat(rand)!: rm init and Random ptr storage
This commit changes the structure and APIs of distribution structs by
removing most of their init methods and no longer storing a pointer to a
Random generator in the structs themselves. Instead, the random
generator is passed as a pointer to the sample/sampleSlice methods.

BREAKING CHANGE: Distribution struct and sampling APIs no longer use a
sotred pointer to a Random struct. Calls to sampling methods must pass
in a Random generator as a function argument.